# Conflicts:
#	processor/src/main/kotlin/com/lightningkite/lightningdb/AnnotationProcessor.kt
#	processor/src/main/kotlin/com/lightningkite/lightningdb/MongoFields.kt
#	server-core/src/main/kotlin/com/lightningkite/lightningserver/auth/subject/AuthEndpointsForSubject.kt
#	server-core/src/main/kotlin/com/lightningkite/lightningserver/files/FileObject.ext.kt
#	server-core/src/test/kotlin/com/lightningkite/lightningserver/jsonschema/SchemaTest.kt
#	server-mongo/src/main/kotlin/com/lightningkite/lightningdb/helpers.kt
#	server-postgresql/src/main/kotlin/com/lightningkite/lightningserver/db/DbMapLikeFormat.kt
#	server-testing/src/main/kotlin/com/lightningkite/lightningdb/test/ConditionTests.kt
#	server-testing/src/test/kotlin/com/lightningkite/lightningserver/TestSettings.kt
#	shared/src/commonMain/kotlin/com/lightningkite/lightningdb/MySealedClassSerializer.kt

+++ server-core/src/main/kotlin/com/lightningkite/lightningserver/serialization/UUIDPartsSerializer.kt
@@ -11,6 +11,7 @@ import kotlinx.serialization.encoding.Decoder
 import kotlinx.serialization.encoding.Encoder
 import kotlinx.datetime.Instant
 import java.util.*
+import com.lightningkite.UUID
 
 object UUIDPartsSerializer: KSerializer<UUID> {
     override val descriptor: SerialDescriptor = buildClassSerialDescriptor("UUID") {
@@ -37,8 +38,8 @@ object UUIDPartsSerializer: KSerializer<UUID> {
 
     override fun serialize(encoder: Encoder, value: UUID) {
         val s = encoder.beginStructure(descriptor)
-        s.encodeLongElement(descriptor, 0, value.mostSignificantBits)
-        s.encodeLongElement(descriptor, 1, value.leastSignificantBits)
+        s.encodeLongElement(descriptor, 0, value.mostSignificantBits.toLong())
+        s.encodeLongElement(descriptor, 1, value.leastSignificantBits.toLong())
         s.endStructure(descriptor)
     }
 }
